+++ src/types/algebraic.rs
@@ -11,6 +11,9 @@ use crate::types::concept::{DataType, Type};
 
 use crate::types::sequence;
 use crate::types::sequence::Tuple;
+use crate::value::concept::ValueCell;
+use crate::value::error::TypeResult;
+use crate::value::union::SumValue;
 
 /// This struct describes a type which can have a value of one of its variant.
 #[derive(Clone, Debug)]
@@ -18,7 +21,7 @@ pub struct SumType(Tuple);
 
 impl SumType {
     pub fn new(types: &[Type]) -> Self { Self(types.to_vec()) }
-    pub fn to_sequence_type(&self) -> Tuple { self.0.to_vec() }
+    pub fn to_tuple(&self) -> Tuple { self.0.to_vec() }
     pub fn variant(&self, tag: usize) -> Option<Type> { self.0.get(tag).cloned() }
     pub fn to_rc(self) -> Rc<Self> { Rc::new(self) }
 }
@@ -30,6 +33,10 @@ impl DataType for SumType {
 
     fn typename(&self) -> String {
         format!("<{}>", sequence::join(&self.0, "|"))
+    }
+
+    fn construct_from_raw(&self, raw: &[u8]) -> TypeResult<ValueCell> {
+        Ok(SumValue::from(self.clone().to_rc(), raw)?.to_cell())
     }
 }
 
@@ -50,7 +57,7 @@ pub struct ProductType(Tuple);
 impl ProductType {
     pub const fn unit_type() -> Self { Self(Vec::new()) }
     pub fn new(types: &[Type]) -> Self { Self(types.to_vec()) }
-    pub fn to_sequence_type(&self) -> Tuple { self.0.to_vec() }
+    pub fn to_tuple(&self) -> Tuple { self.0.to_vec() }
     pub fn is_unit_type(&self) -> bool { self.0.is_empty() }
     pub fn to_rc(self) -> Rc<Self> { Rc::new(self) }
 }
@@ -69,4 +76,8 @@ impl DataType for ProductType {
     fn size(&self) -> usize { self.0.size() }
 
     fn typename(&self) -> String { format!("<{}>", sequence::join(&self.0, "&")) }
+
+    fn construct_from_raw(&self, raw: &[u8]) -> TypeResult<ValueCell> {
+        self.0.construct_from_raw(raw)
+    }
 }