+++ src/types/primitive.rs
@@ -9,6 +9,14 @@
 
 use std::rc::Rc;
 use crate::types::concept::{DataType, Type};
+use crate::value::boolean::Boolean;
+use crate::value::byte::Bytes;
+use crate::value::concept::ValueCell;
+use crate::value::error::{TypeError, TypeResult};
+use crate::value::list;
+use crate::value::number::{Float, Integer};
+use crate::value::reference::Reference;
+use crate::value::list::List as ListValue;
 
 pub type List = (Type, usize);
 
@@ -17,36 +25,37 @@ pub type List = (Type, usize);
 pub enum Primitive {
     /// A type to hold `true` or `false` values.
     ///
-    /// Value is [crate::value::boolean::Boolean]
+    /// Value is [Boolean]
     Boolean,
     /// A type to manage byte at low level.
     ///
-    /// Value is [crate::value::byte::Bytes::Byte]
+    /// Value is [Bytes::Byte]
     Byte,
     /// A type to manage words composed of bytes at low level.
     ///
     /// Value can be :
-    /// * [crate::value::byte::Bytes::Word] for size = 2,
-    /// * [crate::value::byte::Bytes::Quad] for size = 4,
-    /// * [crate::value::byte::Bytes::Long] for size = 8,
-    /// * [crate::value::byte::Bytes::Wide] for size = 16, or,
-    /// * [crate::value::byte::Bytes::Bytes] for any other size.
+    /// * [Bytes::Arch] for size = `size_of::<usize>()`,
+    /// * [Bytes::Word] for size = 2,
+    /// * [Bytes::Quad] for size = 4,
+    /// * [Bytes::Long] for size = 8,
+    /// * [Bytes::Wide] for size = 16, or,
+    /// * [Bytes::Bytes] for any other size.
     Bytes(usize),
     /// A type to perform floating-point arithmetic.
-    /// 
-    /// Value is [crate::value::number::Float]
+    ///
+    /// Value is [Float]
     Float,
     /// A type to perform integral arithmetic.
     ///
-    /// Value is [crate::value::number::Integer]
+    /// Value is [Integer]
     Integer,
     /// A type intended to retain an address to a value of a specific type.
     ///
-    /// Value is [crate::value::reference::Reference]
+    /// Value is [Reference]
     Reference(Type),
     /// A collection type which contains many elements of a specific type.
     ///
-    /// Value is [crate::value::list::List]
+    /// Value is [list::List]
     List(List),
 }
 
@@ -90,6 +99,28 @@ impl DataType for Primitive {
             Primitive::Integer => "integer".to_string(),
             Primitive::Reference(t) => format!("&{}", t.typename()),
             Primitive::List(l) => format!("[{};{}]", l.0.typename(), l.1),
+        }
+    }
+
+    fn construct_from_raw(&self, raw: &[u8]) -> TypeResult<ValueCell> {
+        match self {
+            Primitive::Boolean => Ok(Boolean::from(raw).to_cell()),
+            Primitive::Byte => if raw.len() == 1 { Ok(Bytes::from(raw).to_cell()) } else {
+                Err(TypeError::InvalidType {
+                    expected: self.clone().to_rc(),
+                    provided: Primitive::Bytes(raw.len()).to_rc(),
+                })
+            },
+            Primitive::Bytes(s) => if raw.len() == *s { Ok(Bytes::from(raw).to_cell()) } else {
+                Err(TypeError::InvalidType {
+                    expected: self.clone().to_rc(),
+                    provided: Primitive::Bytes(raw.len()).to_rc(),
+                })
+            },
+            Primitive::Float => Ok(Float::from(raw).to_cell()),
+            Primitive::Integer => Ok(Integer::from(raw).to_cell()),
+            Primitive::Reference(t) => Ok(Reference::from(t.clone(), raw).to_cell()),
+            Primitive::List(t) => Ok(ListValue::from(t.0.clone(), t.1, raw).to_cell())
         }
     }
 }